Palugada Webapp/requirements.txt
Flask==2.2.2
Flask_Bootstrap==3.3.7.1
gTTS==2.2.4
numpy==1.22.4
pandas==1.2.3
Pillow==9.3.0
PyPDF2==2.11.1
gunicorn==20.1.0
